Add clap
Parse command line args with the `clap` crate. Move the current tracing
code into the `painterly` subcommand, which is what it is called in
the code package from TU Wien.

Add more build and run information to the readme.

Using this thing is actually a little bit verbose! I'm excited for v3
to come out.